+++ plugins/CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.1.0-dev.3
+
+- Trim base64 string.
+
 ## 0.1.0-dev.2
 
-- Initial version.
+- ECC utils ok.
 
 ## 0.1.0-dev.1